Add support for switched MAC addresses
This patch allows n2n to properly route MAC address behind the edge nodes.
The map "sw_macs" maps such "switched" MAC address to the edge node to use.
When sending a packet, the destination MAC is first searched in sw_macs.
If a match is found, the destination MAC reported into the N2N packet
metadata is replaced with the edge node MAC. This allows a supernode,
which has no knowledge of the switched MAC, to properly route the packet
to the destination edge.