build(CMakeLists): dynamic link to libcap by default
This is a only exception while all other libraries are dynamically
linked.

The "TODO" does not make any sense as libcap provides both dynamic
and static libraries.

There is not reason to explain why it needs to be statically linked
'especially' as dynamic linking works very well.

Then just make it dynamically-linked by default, which also matches
the behaviour on autotools.

Compiled and tested on aarch64, no errors found.

Signed-off-by: Tianling Shen <i@cnsztl.eu.org>